% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_table.R
\docType{data}
\name{codon_table}
\alias{codon_table}
\title{Codon table.}
\format{
A data frame with 64 rows and 2 variables:
\describe{
\item{codon}{codon, in 3-mers}
\item{amino}{amino acid code, in 1-mer}
...
}
}
\source{
\url{https://github.com/lingminhao/Project-Rosalind/blob/main/Bioinformatics\%20Textbook\%20Track/Chapter\%204/CodonTable.txt}
}
\usage{
codon_table
}
\description{
A dataset containing the maps between codon and its respective amino acid
representation.
}
\keyword{datasets}
