% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcription}
\alias{transcription}
\title{Convert DNA string to RNA string}
\usage{
transcription(dna)
}
\arguments{
\item{dna}{A character vector with one element}
}
\value{
A character vector
}
\description{
Convert DNA string to RNA string
}
\examples{
dna <- "GATGGAACTTGACTACGTAAATT"
transcription(dna)
}
