% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{Convert RNA string to Amino acid string}
\usage{
translation(rna)
}
\arguments{
\item{rna}{A character vector with one element}
}
\value{
A character vector
}
\description{
Convert RNA string to Amino acid string
}
\examples{
rna <- "AUGGCCAUGGCGCCCAGAACUGAGAUCAAUAGUACCCGUAUUAACGGGUGA"
translation(rna)
}
